example, tests: make hostaddr uint32_t
Fixes:
```
example/scp.c:96:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/scp.c:96:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/scp_nonblock.c:150:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/scp_nonblock.c:150:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/scp_write.c:119:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/scp_write.c:119:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/scp_write_nonblock.c:149:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/scp_write_nonblock.c:149:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp.c:160:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp.c:160:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_append.c:114:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_append.c:114:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_mkdir.c:98:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_mkdir.c:98:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_mkdir_nonblock.c:98:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_mkdir_nonblock.c:98:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_nonblock.c:155:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_nonblock.c:155:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_write.c:113:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_write.c:113:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_write_nonblock.c:153:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_write_nonblock.c:153:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftp_write_sliding.c:153:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftp_write_sliding.c:153:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftpdir.c:134:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftpdir.c:134:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/sftpdir_nonblock.c:114:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/sftpdir_nonblock.c:114:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/ssh2.c:128:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/ssh2.c:128:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/ssh2_agent.c:100:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/ssh2_agent.c:100:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/ssh2_agent_forwarding.c:133:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/ssh2_agent_forwarding.c:133:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/ssh2_echo.c:139:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/ssh2_echo.c:139:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
example/ssh2_exec.c:143:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
example/ssh2_exec.c:143:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
tests/openssh_fixture.c:395:27: warning: conversion from 'long unsigned int' to 'in_addr_t' {aka 'unsigned int'} may change value [-Wconversion]
tests/openssh_fixture.c:395:27: warning: implicit conversion loses integer precision: 'unsigned long' to 'in_addr_t' (aka 'unsigned int') [-Wshorten-64-to-32]
```

+++ example/scp.c
@@ -42,7 +42,7 @@
 
 int main(int argc, char *argv[])
 {
-    unsigned long hostaddr;
+    uint32_t hostaddr;
     libssh2_socket_t sock;
     int i, auth_pw = 1;
     struct sockaddr_in sin;
